Alow definition of spi device in device tree.
This patch will use module parameters if they are defined, otherwise
it will use the device tree. Using the device tree is preferred as
we can use the same driver for other boards.

This method is not only preferred, but also required as
spi_busnum_to_master() is not available and alternative methods
are discouraged and not available in future kernels.

Signed-off-by: Jesse Taube <Mr.Bossman075@gmail.com>